pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "coachai"
version = "0.1.0"
description = "CoachAI - 智能伴读AI系统"
authors = ["baofengbaofeng <baofengbaofeng@example.com>"]
license = "GPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/baofengbaofeng/coach-ai"
repository = "https://github.com/baofengbaofeng/coach-ai"
keywords = ["education", "ai", "homework", "exercise", "learning"]

[tool.poetry.dependencies]
python = "^3.14"
Django = "^5.0.2"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
django-filter = "^23.5"
mysqlclient = "^2.2.4"
django-environ = "^0.11.2"
channels = "^4.0.0"
channels-redis = "^4.1.0"
daphne = "^4.0.0"
paddleocr = "^2.7.0.3"
easyocr = "^1.7.1"
opencv-python = "^4.9.0.80"
mediapipe = "^0.10.8"
openai-whisper = "^20231117"
SpeechRecognition = "^3.10.1"
Pillow = "^10.2.0"
python-magic = "^0.4.27"
celery = "^5.3.6"
redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
flake8 = "^7.0.0"
pytest = "^7.4.4"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
locust = "^2.20.2"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
coachai = "coachai.manage:main"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 120
extend-ignore = "E203,W503"
exclude = [
  ".git",
  "__pycache__",
  ".venv",
  "venv",
  "*.pyc",
  "*.pyo",
  "*.pyd",
  ".Python",
  "build",
  "develop-eggs",
  "dist",
  "downloads",
  "eggs",
  ".eggs",
  "lib",
  "lib64",
  "parts",
  "sdist",
  "var",
  "wheels",
  "*.egg-info",
  ".installed.cfg",
  "*.egg",
  "MANIFEST",
  ".pytest_cache",
  ".mypy_cache",
  ".tox",
  ".coverage",
  "htmlcov",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["tests.py", "test_*.py"]
python_classes = ["*Test"]
python_functions = ["test_*"]
addopts = [
  "--strict-markers",
  "--strict-config",
  "--tb=short",
  "--cov=.",
  "--cov-report=term-missing",
  "--cov-report=html",
]
testpaths = ["tests"]

[tool.coverage.run]
source = ["."]
omit = [
  "*/tests/*",
  "*/migrations/*",
  "*/__pycache__/*",
  "*/settings/*",
  "venv/*",
  ".venv/*",
  "env/*",
  "ENV/*",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug:",
  "if settings.DEBUG",
  "raise AssertionError",
  "raise NotImplementedError",
  "if 0:",
  "if __name__ == .__main__.:",
  "class .*\bProtocol\b",
  "@(abc\\.)?abstractmethod",
]
